Removed the route for edit
